% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{crownExpansion}
\alias{crownExpansion}
\title{Data: Crown wood expansion models to include branch coarse wood}
\format{
a list with two elemens, each a (GAM) model
}
\source{
based on comparison between BDAT and TapeS
}
\usage{
crownExpansion
}
\description{
This data gives percentage of branch coarse wood (as difference between TapeS
and BDAT taper curves) with respect to BDAT crown coarse wood. With these
models one can expand the TapeS main axis crown coarse wood into total
crown coarse wood. Yet, a clear volume reference of very large trees - the
case, where expansion seems to be necessary - does not exists.
}
\examples{
data(crownExpansion)
str(crownExpansion, max.level=1)
}
\references{
none, yet.
}
\keyword{datasets}
