% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{unvd}
\alias{unvd}
\title{Data: Percentage of unusable wood in Beech and Oak in Germany}
\format{
A three-dimensional array.
}
\source{
BDAT-Fortran-library
}
\usage{
unvd
}
\description{
This data gives percentage of unusable wood according to species (beech or
oak) and 2cm-diameter classes and cutting diameter. Data was taken from
BDAT-Fortran-library.
}
\examples{
data(unvd)
dim(unvd)
dbh <- 40 # diameter in breast height
cd <- 10 # cutting diameter
unvd[1, floor(dbh/2), cd]
}
\references{
none, yet.
}
\keyword{datasets}
